Minorest, utils/easing/easing.cpp - fixed some double to float warnings by changing pow to powf

+++ source/utils/easing/easing.cpp
@@ -110,7 +110,7 @@ float Elastic::impl_easeIn (float t,float b , float c, float d) {
 	float p=d*.3f;
 	float a=c; 
 	float s=p/4;
-	float postFix =a*pow(2,10*(t-=1)); // this is a fix, again, with post-increment operators
+	float postFix =a*powf(2,10*(t-=1)); // this is a fix, again, with post-increment operators
 	return (float) -(postFix * sin((t*d-s)*(2*PI)/p )) + b;
 }
 
@@ -129,10 +129,10 @@ float Elastic::impl_easeInOut(float t,float b , float c, float d) {
 	float s=p/4;
 	 
 	if (t < 1) {
-		float postFix =a*pow(2,10*(t-=1)); // postIncrement is evil
+		float postFix =a*powf(2,10*(t-=1)); // postIncrement is evil
 		return (float)( -.5f*(postFix* sin( (t*d-s)*(2*PI)/p )) + b );
 	} 
-	float postFix =  a*pow(2,-10*(t-=1)); // postIncrement is evil
+	float postFix =  a*powf(2,-10*(t-=1)); // postIncrement is evil
 	return (float)( postFix * sin( (t*d-s)*(2*PI)/p )*.5f + c + b );
 }
 
@@ -143,17 +143,17 @@ Expo::EaseOut Expo::easeOut;
 Expo::EaseInOut Expo::easeInOut;
 
 float Expo::impl_easeIn (float t,float b , float c, float d) {
-	return (t==0) ? b : c * pow(2, 10 * (t/d - 1)) + b;
+	return (t==0) ? b : c * powf(2, 10 * (t/d - 1)) + b;
 }
 float Expo::impl_easeOut(float t,float b , float c, float d) {
-	return (t==d) ? b+c : c * (-pow(2, -10 * t/d) + 1) + b;	
+	return (t==d) ? b+c : c * (-powf(2, -10 * t/d) + 1) + b;	
 }
 
 float Expo::impl_easeInOut(float t,float b , float c, float d) {
 	if (t==0) return b;
 	if (t==d) return b+c;
-	if ((t/=d/2) < 1) return c/2 * pow(2, 10 * (t - 1)) + b;
-	return c/2 * (-pow(2, -10 * --t) + 2) + b;
+	if ((t/=d/2) < 1) return c/2 * powf(2, 10 * (t - 1)) + b;
+	return c/2 * (-powf(2, -10 * --t) + 2) + b;
 }
 
 //-----------------------------------------------------------------------------
